Move include to spauly and rename namespace
